Add Grouping to Single Cell Data

The add_grouping_single_cell function iterates over a list of cell types and assigns labels to each single cell in a dataset based on its group at each level. The labels are in the format 'subcluster_level_#leaf#'.

add_grouping_single_cell(
  single_cell_data_exp,
  sub_clusters_var,
  level_group_name,
  hierarchy_level_groups
)

Arguments

single_cell_data_exp

A list containing the single cell data. The function will add new labels to this list.

sub_clusters_var

A string representing the variable used to indicate the subclusters in single_cell_data_exp.

level_group_name

A string representing the name of the level group variable in single_cell_data_exp. The function will add labels to this variable.

hierarchy_level_groups

A list of cell types that define the hierarchy levels. Each cell type should be a list itself with the elements celltype_list, tree_level, and leaf_number.

Value

The function returns the single_cell_data_exp list with added labels to the level_group_name variable.