Calculate the Best Global Cluster

Creates the tree of cluster groups and levels. Decides the number of clusters that will minimize the correlation between them, given a list of cell types.

calculate_best_cluster_global(
  single_cell_data_exp,
  var_cell_type,
  subset_celltypes = NULL,
  var_sample,
  use_min_cor_strategy = TRUE,
  delete_shared_level_markers = FALSE,
  delete_shared_internal_markers = FALSE,
  number_clusters_one_celltype = 1,
  distance_method = "euclidean",
  hclust_method = "average",
  min_size_leaf = 3,
  max_k_tried_hier_clustering = 3,
  random_seed = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  filter_main_markers = TRUE,
  verbose = FALSE
)

Arguments

single_cell_data_exp

A data frame or matrix containing single-cell expression data.

var_cell_type

A vector or factor indicating the cell type for each cell.

subset_celltypes

A vector of cell types to subset for the analysis. Default is NULL, implying all cell types are used.

var_sample

A vector or factor indicating the sample source for each cell.

use_min_cor_strategy

A logical value indicating whether to use the minimum correlation strategy. Default is TRUE.

delete_shared_level_markers

A logical value indicating whether to delete markers shared at level of the hierarchy. Default is FALSE.

delete_shared_internal_markers

A logical value indicating whether to delete markers shared internally. Default is FALSE.

number_clusters_one_celltype

An integer specifying the number of clusters for each cell type. Default is 1.

distance_method

A string specifying the method to compute the distance. Default is 'euclidean'.

hclust_method

A string specifying the method to perform hierarchical clustering. Default is 'average'.

min_size_leaf

An integer specifying the minimum size of the leaf in the hierarchical clustering. Default is 3.

max_k_tried_hier_clustering

An integer specifying the maximum number of clusters tried in hierarchical clustering. Default is 3.

random_seed

An integer specifying the random seed for reproducibility. Default is NULL.

param.logfc.threshold

A numeric value indicating the log-fold-change threshold for marker selection. Default is 2.0.

param.p_val_adj

A numeric value indicating the adjusted p-value threshold for marker selection. Default is 0.05.

filter_main_markers

A logical value indicating whether to filter main markers. Default is TRUE.

verbose

A logical value indicating whether to print additional output during the computation. Default is FALSE.

Value

A list containing results of the clustering process.