Calculate Optimal Clustering at Generic Level

This function recursively analyses a level of clustering by attempting to separate it into 2, 3, or 4 clusters. The aim is to minimize the correlation between the groups. For each subgroup, the function is called again until each cluster contains 2 or 3 cell types. This enables granular and optimized clustering across different levels of hierarchy.

calculate_best_clustering_generic(
  single_cell_data_exp,
  reference,
  var_cell_type,
  cell_types,
  var_sample,
  tree_level = 0,
  leaf_number = 1,
  number_clusters_one_celltype = 1,
  distance_method = "euclidean",
  hclust_method = "average",
  min_size_leaf = 3,
  max_k_tried_hier_clustering = 3,
  verbose = FALSE
)

Arguments

single_cell_data_exp

A data frame or matrix containing single-cell expression data.

reference

A reference data set or object to guide the clustering process.

var_cell_type

A string or variable indicating the cell type for each cell.

cell_types

A vector of cell types present in the data.

var_sample

A string or variable indicating the sample for each cell.

tree_level

An integer indicating the current level of the tree hierarchy in the clustering process. Default is 0.

leaf_number

An integer indicating the number of leaf nodes in the tree hierarchy. Default is 1.

number_clusters_one_celltype

An integer indicating the number of clusters per cell type. Default is 1.

distance_method

A string indicating the distance measure to be used for clustering. Default is 'euclidean'.

hclust_method

A string indicating the method to be used for hierarchical clustering. Default is 'average'.

min_size_leaf

An integer indicating the minimum size of the leaf nodes. Default is 3.

max_k_tried_hier_clustering

An integer indicating the maximum number of clusters tried in the hierarchical clustering. Default is 3.

verbose

A logical value indicating whether to print additional output during the computation. Default is FALSE.

Value

A list containing the optimal clustering results.