Plot Performance of Multiple Deconvolutions

This function generates a plot showing the performance of multiple deconvolutions. The plot visualizes the divergence and delta divergence metrics against a main parameter of interest. Besides it is used for plotting the results for the gridSearch in a general level.

performance_plot_multiple_deconvolutions(
  performance_df,
  main_parameter = "alpha"
)

Arguments

performance_df

A data frame that contains performance metrics for each deconvolution result. The data frame should include the main parameter, divergence, and delta divergence values.

main_parameter

A string that specifies the name of the main parameter to be plotted on the x-axis.This parameter should be a column in the 'performance_df' data frame.

Value

A ggplot object showing the performance of multiple deconvolutions against the specified main parameter.

Details

The function uses ggplot2 to generate a line plot where the y-axis represents the logarithmic values of the divergence and delta divergence metrics, and the x-axis represents the main parameter. The divergence and delta divergence metrics are differentiated by color.