Plot Deconvolution Performance Metrics

This function plots various performance metrics for a deconvolution process across multiple iterations. Function for gridSearch analysis: General plots.

plot_deconvolution_performance(running_info, title_addition = "")

Arguments

running_info

A data frame with iteration number and the corresponding values for divergence and delta divergence.

title_addition

An optional string that is added to the plot title.

Value

The function returns a list of ggplot objects, each representing a different visualization of the deconvolution performance.

Details

This function creates three plots to visualize the performance of a deconvolution process:

The natural logarithm of the divergence value over iterations.

The natural logarithm of the delta divergence value over iterations.

Both the above metrics in the same plot. The divergence value provides a measure of how much the deconvolution results deviate from the ground truth, while the delta divergence gives a measure of how much the divergence changes between successive iterations.