Plot Performance for Each Model

This function plots the performance of each model from a grid search result based on specified alpha and beta values.

plot_performance_for_each_model(
  grid_search_result,
  alpha = NULL,
  beta = NULL,
  plot_type = "summary",
  scale_parameters = FALSE,
  save.pdf = FALSE,
  path.pdf = NULL
)

Arguments

grid_search_result

A list object that contains the results of a grid search.

alpha

Numeric, the alpha value to filter the models to plot. Optional, default is NULL.

beta

Numeric, the beta value to filter the models to plot. Optional, default is NULL.

plot_type

Character, specifies the type of plot to produce. Options are 'summary' (default) or other types supported by the underlying plotting function.

scale_parameters

Logical, whether to scale the parameters in the plot. Optional, default is FALSE.

save.pdf

Logical, whether to save the plot as a PDF. Optional, default is FALSE.

path.pdf

String, the path where to save the PDF file if 'save.pdf' is TRUE. Optional, default is NULL.

Value

The function is used for its side effects of creating and optionally saving plots. It does not return anything.

Details

The function iterates over each model in the grid search result. For each model, it produces a plot based on the 'plot_type' argument. It filters the models to plot based on the 'alpha' and 'beta' arguments.

If 'save.pdf' is TRUE, it saves the plot as a PDF file to the specified path or the current working directory if no path is specified.