Recalculate Cell Type Proportions

This function recalculates the proportions of cell types based on the top-level clustering proportions and the more granular sub-cluster proportions.

recalculate_proportions(
  top_cluster_deco_list,
  top_clusters_list,
  subcluster_deco_list,
  verbose = FALSE
)

Arguments

top_cluster_deco_list

A list containing the deconvolution results for the top-level clusters.

top_clusters_list

A vector containing the names or identifiers of the top-level clusters.

subcluster_deco_list

A list of deconvolution results for the sub-clusters.

verbose

A logical indicating whether or not to print detailed messages while the function is running. Default is FALSE.

Value

A matrix where each row represents a cell type and each column represents a sample. Each entry in the matrix represents the recalculated proportion of a cell type in a sample.

Details

This function takes as input the top-level cluster proportions and a list of sub-cluster deconvolution results. It scales the sub-cluster proportions by the corresponding top-level cluster proportions and then combines all the proportions. This results in an updated set of cell type proportions that reflects the hierarchical structure of the clustering.