run_deconvolution_simulation_generic_recursive

Perform a recursive deconvolution simulation on given single cell data. This function handles the first top node of the hierarchy and then send the rest to the recursive algorithm

run_deconvolution_simulation_generic_recursive(
  single_cell_data_exp,
  hierarchy,
  sub_clusters_var,
  sample,
  use_min_cor_strategy = TRUE,
  delete_shared_level_markers = FALSE,
  delete_shared_internal_markers = FALSE,
  deconvolute_just_top = FALSE,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  marker_strategy = "keep_default_values",
  verbose = FALSE
)

Arguments

single_cell_data_exp

A data frame that contains single cell expression data.

hierarchy

A list that outlines the hierarchy of cell types.

sub_clusters_var

A variable that represents the subclusters in the data.

sample

A variable that represents the samples in the data.

use_min_cor_strategy

A logical parameter indicating whether to use minimum correlation strategy. Defaults to TRUE.

delete_shared_level_markers

A logical parameter indicating whether to delete shared level markers. Defaults to FALSE.

delete_shared_internal_markers

A logical parameter indicating whether to delete shared internal markers. Defaults to FALSE.

deconvolute_just_top

A logical parameter indicating whether to only deconvolute top. Defaults to FALSE.

filter_markers

A parameter for specifying markers to filter, defaults to NULL.

param.logfc.threshold

A numeric value specifying the threshold for log fold change, defaults to 2.0.

param.p_val_adj

A numeric value specifying the threshold for adjusted p-values, defaults to 0.05.

marker_strategy

Strategy for selecting markers, defaults to NULL.

verbose

A logical parameter indicating whether to print detailed messages during the execution of the function. Defaults to FALSE.

Value

A list containing the deconvolution result and the pseudo bulk data.