Tree-guided Deconvolution of Bulk Data

run_deconvolution_tree_guided is a generic function for deconvoluting bulk data based on a reference single-cell dataset using a tree-guided method.

run_deconvolution_tree_guided(
  bulk_data,
  single_cell_data_exp,
  top_clusters_var,
  top_clusters_list,
  sub_clusters_var,
  sub_clusters_list,
  sample,
  use_min_cor_strategy = FALSE,
  true_proportions = NULL,
  delete_shared_level_markers = TRUE,
  delete_shared_internal_markers = TRUE,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  test.use.value = "wilcox",
  marker_strategy = "keep_default_values",
  minimum_markers = 4,
  min_delta_cor_threshold = 0.05,
  verbose = FALSE
)

Arguments

bulk_data

A numeric matrix representing the bulk data to be deconvoluted.

single_cell_data_exp

A list or data.frame representing the single cell data.

top_clusters_var

A string representing the variable in single_cell_data_exp used to indicate the top clusters.

top_clusters_list

A list or vector indicating the top clusters to be considered.

sub_clusters_var

A string representing the variable in single_cell_data_exp used to indicate the subclusters.

sub_clusters_list

A list or vector indicating the subclusters to be considered.

sample

A string or numeric value indicating the specific sample to use.

use_min_cor_strategy

A logical value indicating whether to use the minimal correlation strategy. Default is FALSE.

true_proportions

A numeric vector representing the true proportions of the clusters, if known. Default is NULL.

delete_shared_level_markers

A logical value indicating whether to delete shared markers at the same level. Default is TRUE.

delete_shared_internal_markers

A logical value indicating whether to delete shared markers within a group. Default is TRUE.

filter_markers

A character string specifying the filtering strategy for markers. Default is NULL.

param.logfc.threshold

A numeric value representing the log fold change threshold for marker genes. Default is 2.0.

param.p_val_adj

A numeric value representing the adjusted p-value threshold for marker genes. Default is 0.05.

test.use.value

A string specifying the test to use for marker gene identification. Default is 'wilcox'.

marker_strategy

A character string specifying the overall strategy for marker gene handling. Default is NULL.

minimum_markers

The minimum number of marker genes to be selected, default is set to 4. This can be adjusted based on the analysis requirements.

min_delta_cor_threshold

Minimum difference between the correlation of t and t-1 steps. The idea is to skip genes that have small changes in the correlation. For default it is a 5%.

verbose

A logical value indicating whether to print detailed output during execution. Default is FALSE.

Value

A list containing the deconvoluted bulk data.