Save Plot as PDF

This function saves a given plot object as a PDF file to a specific path.

save_plot_as_pdf(
  plot_object,
  file_name,
  width.parameter = NULL,
  height.parameter = NULL,
  path = NULL
)

Arguments

plot_object

The plot object to be saved.

file_name

String, the name of the file to save the plot as. It should not include the path or file extension.

width.parameter

Numeric, the width of the output PDF in inches. Optional, default is NULL.

height.parameter

Numeric, the height of the output PDF in inches. Optional, default is NULL.

path

String, the directory where the PDF should be saved. Optional, default is '/mnt_volumen/GIT_REPOSITORIES/MTF/results/'.

Value

The function is used for its side effects of creating a PDF file and does not return anything.

Details

The function first checks whether a path was provided. If not, it uses a default path. Then, it opens a new PDF device, prints the 'plot_object' to this device, and then closes the device, which saves the plot to the specified file. The 'width.parameter' and 'height.parameter' arguments allow for control over the dimensions of the output PDF. If these are not provided, the PDF is created with the default dimensions.